Avoid requiring changes to API (OpenTelemetry::Adapter)
* avoid making API changes now by inlining required methods
  (#propagator)
